backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pymongo==4.11
python-dotenv==1.0.1